Rename `api` module to `syntax`
