+++ crates/rome_js_syntax/src/lib.rs
@@ -17,8 +17,9 @@ pub use self::generated::*;
 pub use ast::{AstNode, AstNodeList, AstSeparatedList, AstToken, SyntaxError, SyntaxResult};
 pub use expr_ext::*;
 pub use modifier_ext::*;
-use rome_rowan::api::TriviaPieceKind;
-pub use rome_rowan::{SyntaxText, TextLen, TextRange, TextSize, TokenAtOffset, WalkEvent};
+pub use rome_rowan::{
+    SyntaxText, TextLen, TextRange, TextSize, TokenAtOffset, TriviaPieceKind, WalkEvent,
+};
 pub use stmt_ext::*;
 pub use syntax_node::*;
 pub use util::SyntaxNodeExt;
